+++ tests/test_import_dbml.py
@@ -13,7 +13,6 @@ def test_cli():
         app,
         [
             "import",
-            "--format",
             "dbml",
             "--source",
             "fixtures/dbml/import/dbml.txt",
@@ -28,13 +27,12 @@ def test_cli_with_filters():
         app,
         [
             "import",
-            "--format",
             "dbml",
             "--source",
             "fixtures/dbml/import/dbml.txt",
-            "--dbml-schema",
+            "--schema",
             "test",
-            "--dbml-table",
+            "--table",
             "foo",
         ],
     )